Figure 3
Figure 3. Pooled sequencing revealed a strong candidate locus (single-minded 1, SIM1) that may underly the slender anole dewlap polymorphism. (A) Manhattan plot illustrating genomic differentiation between the dewlap morphs using pairwise FST values, and (B) Fisher Exact Tests. The red lines in (B) and (C) represent the Bonferroni Correction (p-value = 8.9), with points above the line indicating significantly differentiated SNPs across the slender anole genome. (C) One locus on scaffold 3, the transcription factor SIM1, contained a peak of many highly differentiated SNPs between morphs. The vertical green bars in the inset represent exons and the arrow represents the direction of genome annotation. (D) Nucleotide diversity (𝝅) of SIM1 for the bicolor (green) and solid (brown) dewlap morphs. Note that nucleotide diversity was often higher in the dominant solid morph. (E) SNP panel showing the major allele frequencies of all significant SNPs (n = 175) in the candidate region of SIM1 for each morph plotted as heatmaps. We denote the allele that is either homozygous or heterozygous in the dominant solid morph as the reference “ref” and the allele that is normally homozygous in the recessive bicolor morph as the alternative “alt”). Note that most SNPs are heterozygous in the solid morph but are usually homozygous in the bicolor morph. (F) Male slender anole dewlaps come in two morphs: solid and bicolor.